requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
entrypoints==0.3
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyxlsb==1.0.6
pyzmq==19.0.1
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
xlrd==1.2.0
zipp==3.1.0
